LCOW: Add environment variable to enable
